Add files via upload
Some of the assets powering the visual element of the game have been extracted from https://github.com/KelvinThai/floppyBird_gamefi/tree/main/game.
Feel free to modify the game and use it for your personal use. If any changes to be made please raise an issue.

If the copyright holder would like for the assets to be removed, please open an issue to start the conversation.